Fix PlatformInputContext::setFocusObject for editable areas
PlatformInputContext::setFocusObject assumes that if a focus object
rejects input methods once, it will never need them until the object
stays the same.
This is not the case for QtWebEngine where a web page might have
editable areas like input fields and non-editable areas on the same
QQuickItem. If the user touches the non-editable area first the virtual
keyboard will never query the object again, thus the events will not
be delivered to the input field, even if the user touches an editable
area on the page later.
Make the input context always query whether the focus object accepts
input method events regardless of the object being the same.

Change-Id: I943227195a5355f380e684ad5a5d889cabf611f0
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>