Update the current input locale according to layout
The input locale is now updated (temporarily) if the keyboard layout
requires. For example when entering into phone number layout (which
supports only English layout), the input locale is switched into en_GB.
The keyboard stores the actual input locale in different variable and
switches the input locale to that after when moving on to the next
input item.

Improved logic for switching between the available keyboard locales.
The keyboard now remembers the current input locale when moving into
a input field that has limited options for different layouts. For
example in digits layout, where the keyboard offers only Arabic and
English layout, the available input locales are resolved into "Current
locale" and Arabic locale, where the "Current locale" can be any input
locale that shares the English digits layout.

Change-Id: I92dac9e81a8dbdacb1a65784f8a0e51621d9aa67
Reviewed-by: Rainer Keller <rainer.keller@digia.com>