Allow pressing a key before a previous key was released
The current behavior is that the seconc key is completely ignored
since we are using a MouseArea, which can only track one touch point
at a time.  The problem is encountered often while typing with both
hands.

Fix the issue by using a MultiPointTouchArea instead, and track which
of the pressed touch points is active (always the last pressed one).
If a finger is pressed before a previous finger was released, assume
that the user wanted to press that key and register the input method
release before the finger was physically released. This matches the
behavior of the virtual keyboard on my Android device.

MultiPointTouchArea emulates mouse events through a touch point and
most of the behavior is the same. PressAndHold isn't handled for us
so we have to use a Timer to handle it explicitly.

Change-Id: Id685364fc439b93089e6748a86c98d51ebd7eb12
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>