Improve input mode handling
The current input mode is now preserved between the layout change,
unless the layout explicitly tell it requires a specific input mode.

There is also an explicit latinOnly mode which enforces PlainInputMethod
and English keyboard layout. In this mode the user cannot change the
input language. This mode does not affect to primary input language.
The primary input language is activated again when the user selects
another input field which does not enforce the latinOnly mode.

Change-Id: I7a7d292e4685ac144dc0f28cadfcd4ef985ddd97
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>