Add arrow navigation mode
The arrow navigation mode allows using the keyboard with the arrow keys
and the return key of the physical input device. The navigation mode is
available when the keyboard is built with CONFIG+=arrow-key-navigation
in the qmake command line.

To enter into arrow navigation mode, simply press any of the arrow keys
on the hardware keyboard while the virtual keyboard is visible. Items
are selected with return key. Pressing and holding the return key will
access the alternative keys (if any). The "drag symbol mode" is also
available by pressing and holding the return on top of the symbol key.

Change-Id: Ief9acfacd7f2d8a59a3c6b959d28f5467d207fa6
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>