Changes to shift state handling for certain languages
Moved all shift handling logic into DeclarativeShiftHandler. The
purpose of this change is to allow language specific exceptions to
the shift handling logic.

With certain languages (currently including Arabic, Persian and
Hindi), the automatic shift handling is disabled. Also, the shift
state can be toggled even if current input method hints does
not allow it.

This change frees the shift key state for other purposes, such as
binding to an alternative layout.

Change-Id: I1cd6f8684b461ca3eecb484ae307c7bc7c2d0d81
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>