Decouple the input method state change from setFocusObject
This fixes issues with QtWebEngine where input fields in web pages
ended up losing input method focus because the IM state update relied
on the PlatformInputContext::setFocusObject call, which for a WebView
only happens once the WebView gains active focus. Consequently if the
user did not click the input field on the web page first, the WebView
reported not to accept input method events and the VKB remained stuck
in this state even if the user clicked the input field since the focus
object did not change.
This patch also updates the IM state in PlatformInputContext::update
which is called by QInputMethod::update every time the IM is triggered
thus we will never miss input method events in the above described
scenario.

Change-Id: I7680b1ef81b2a8b75f4558cf21214754e549e5dd
Reviewed-by: Andras Becsi <andras.becsi@digia.com>
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>