Add two way navigation mode
Improved the navigation logic for the arrow-key-navigation mode.

The navigation mode can now be used with left+right+enter keys only.
When navigating horizontally to the edge or the keyboard, the
navigation focus moves to the next or previous row. The word selection
list is focused when navigating out of the keyboard from the first or
the last row.

Also, improved logic for navigating away from the word selection list
with up/down keys. Instead of returning to the previous highlighted
key, the navigation focus is either moved to key on first row (down)
or last row (up).

Change-Id: Ie431fad8dda975979b40cc882940cf3a97eb9dd3
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>