Optimize loading time of the keyboard with sound effects
It seems that by using the Component instead of Qt.createQmlObject
reduces the loading time of an single SoundEffect instance to fraction.

Initial analysis shows that the loading time of an keyboard layout
is reduced to half on particular system. Tested by running the
virtual keyboard with basic widgets application in QML Profiler.

Change-Id: I9394ba2d10ee302afcfa33655e4fae94a9029eeb
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>