Add automatic tests
This change introduces an automated testcase which covers the most
essential functionality of the virtual keyboard.

- Input method hints
- Key preview popup
- Alternative keys
- Locale change
- Shift handling (Shift case, lower case, UPPER CASE)
- Symbol mode layout
- Key release inaccuracy handling
- Drag symbol mode
- Theme change
- Sound effects
- Arrow key navigation

Some tests are expected to fail if the keyboard is not compiled
with expected qmake configuration flags (e.g. arrow-key-navigation).

Also, if the keyboard is compiled with pinyin support, all the
tests which expect the international keyboard layout support,
are expected to fail.

Change-Id: Ie23b6628ddceceddd4cb26e70ad382baa8797aaf
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>