Add support for Windows Desktop
Renamed XcbInputPanel component to DesktopInputPanel. The new panel
component provides unified support for both Xcb and Windows desktops.

In ideal case there would be no need for platform specific code, but
unfortunately that is not the case. For example in Xcb, using the
QWindow::setMask combined with Qt::BypassWindowManagerHint does not
work correctly. More complex regions (than a single rectangle) are
not masked correctly. For this reason, the Xcb integration still uses
the Xcb specific code to mask the transparent areas of the input panel.

Changes to PlatformInputContext event filtering are needed because
QPlatformInputContext::filterEvent() method is not functional in
Windows. In this case the QObject::eventFilter() is used instead.
Without these changes the arrow key navigation feature will not work.

The desktop integration can now be disabled with disable-desktop flag
instead of disable-xcb flag. While the disable-xcb still works, it
is deprecated and will be removed in future releases.

Please note that this change adds the Windows Desktop support for the
core keyboard only. The complete Windows support including the 3rdparty
libraries is added in other changes.

Change-Id: Ib65118f715c86746fc7d12ea6d9a2e7ef6e15b09
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>