Reset sound effects when theme is changed
Previously when a theme was changed runtime, the sound effects
remained in the "cache". In this change the old sound effects
are removed before the new ones are loaded.

Also, added signal that is fired when sound effects are played or
stopped. This signal will be used for testing purposes only.

Change-Id: I31defd34400a7c42b1c5eebd4725f35a1ebb36c9
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>