Add HangulInputMethod for Korean language
This change adds support for Korean language. The keyboard layout
is based on the standard Dubeolsik keyboard layout. The keyboard
layout uses the Hangul Compatibility Jamo 3130-318F as the input
characters.

The Hangul composition and decomposition is implemented in the
Hangul support class, and has the following key features:
 - Automatic detection of initial, medial and final Jamos
   based on the current syllable
 - Handles any Unicode input sequences, not just Hangul
 - Splitting and joining of double medial and double final Jamos
 - Splitting of double final consonant and joining with following
   vowel

Added test cases that should cover most of the special cases
in Hangul input.

Change-Id: I5a6870407097bd1ce79bd2d00f81a30bd35a2f91
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>