Add support for pattern recognition based input methods
This change adds generic support for pattern recognition
based input methods.

Added new API for the input engine and input method to
process trace data. The trace data can originate from
various input devices, e.g. from touch screen or from
a dedicated hardware touch panel.

Added new data model type for trace supporting both the C++
and QML interfaces. The new data model is DeclarativeTrace
and Trace respectively, and it stores the trace data captured
from the input device.

The trace object is owned by the input method and is accessible
to the UI layer, capture device and the input method.

First, when the trace event begins, the capture device invokes
the InputEngine.traceBegin(). The input engine forwards this
call to the input method, which creates the trace object in
response to successful call.

Then the capture device receives the trace object and starts
collecting the data. Also, in case of touch screen input, there
are also the UI layer which renders the data. For this purpose
there are new kinds of Style elements available in the Styles
plugin. TraceCanvas is a specialized Canvas for rendering the
trace object on screen. The TraceCanvas is a normal styling
component, and can be customized like any other style element.

Finally, the InputMethod.traceEnd() is called when the trace
interaction ends. The trace is removed from screen automatically
when the trace object is deleted. I.e., the input method has
full control on how many traces it wants to collect for single
recognition phase.

Change-Id: I80ed90032f715726280197d9e94e7f0bd8280ff3
Reviewed-by: Gatis Paeglis <gatis.paeglis@theqtcompany.com>