Fix rules for automatic space insertion
This change applies to automatic space insertion rules in
HunspellInputMethod.

The difference to the existing behavior is that space is not
added when the contents of the editor changes externally (cursor
position, selection, textual content).

This change fixes an issue with replacing text selection with
new composed word. Previously the text was correctly replaced,
but the composition did not start with the first letter. The
reason for this bug was an additional (unexpected) call to
update() function, which terminated the newly started text
composition.

Also, added automatic tests to make sure these rules are not
broken by the future changes.

Task-number: QTRD-3316
Change-Id: If4e90b85b1c3c5d1d98a4e7799a7c24d656cab20
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>