Optimize the selection list updates in PinyinInputMethod
This issue was discovered when experimenting with the
selection list add and remove transitions.

The issue was that in PinyinInputMethod the candidate
list is updated in some cases twice or more where only
one update call is needed.

Added a helper class which updates the candidate list
automatically by monitoring the total candidate count
and the existing candidate list.

Change-Id: I0abbe01198ccce99f33cbdf20c6b5bfd1d063815
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>