Remove manual call to HunspellInputMethod::reset()
The reset() function should not be called directly from the input
method because it is a virtual function and may cause side effects
in the derived class.

This change moves the reset() function to private class and uses
that function instead.

This change does not introduce any new behavior, nor it should break
any existing ones. It fixes a particular issue that presents
itself after introduction of the LipiInputMethod, which uses
HunspellInputMethod as its base class.

Change-Id: I38b56150a677f6c8fcf10a9b7b556acc3f7f4a92
Reviewed-by: Gatis Paeglis <gatis.paeglis@theqtcompany.com>