Reset InputEngine.activeKey if the keyboard.activeKey is reset
This change ensures the activeKey in the input engine is released
if the keyboard layout changes while still pressing the key. This can
happen if a repeating key (for example, the delete) is held using one
finger and the focus is changed to an input field which uses different
kind of keyboard layout using other finger.

Change-Id: Ib923d57cc60256670191af1fbd4f1cecb333486c
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>