Move Hunspell dictionary loading to worker thread
This change moves the Hunspell dictionary loading to worker thread.

The dictionary loading is relatively heavy task for large dictionary
files, such as Arabic, which can take more than 2 seconds to load on
the Nexus 7 hardware.

The Hunspell worker thread is no longer destroyed and created every
time the dictionary changes, instead the worker thread is re-used.

The dictionary loading happens by adding a new kind of task for the
worker thread. Also, added a special function for clearing all but
dictionary loading tasks from the worker task queue. This is needed
for a special case where the dictionary is still loading while the
user starts typing.

Updated the test cases to take into account changes in delays in
loading the dictionary.

Change-Id: If9be14c7703b39af79f6e3b6708e297a28c49f2f
Reviewed-by: Gatis Paeglis <gatis.paeglis@theqtcompany.com>