lipi-toolkit: Fix an infinite loop in NNShapeRecognizer
This change fixes an infinite loop which occurs with the default
alphanumeric model and with a specific subset of classes to be
recognized.

If the subset of classes is set to numbers only, the code path will
end up in an infinite loop.

The issue was that the iterator which caused the infinite loop was
incremented _after_ checking for its end boundary. The fix is to add
the check after the increment, which obviously was the intent of the
original author.

Change-Id: I5821c142018ddd9f301771e34b212e1cb907444b
Reviewed-by: Gatis Paeglis <gatis.paeglis@theqtcompany.com>