Improve input mode handling
Replaced the generic binding for inputMethodHints with
a specific binding to ImhPreferNumbers hint.

The problem with the generic binding was that it occurs
at point which the input mode has already changed, thus
causes an extra change to the input mode, if the numbers
hint is set or reset.

This may be a problem for the specific input methods, which
may load different data model for different input modes
(e.g. handwriting recognition).

Simplified the input mode selection in updateInputMethod()
function. The order and preference which input mode is used
is more clear now.

Note: Updated the Japanese tests because the input mode
is preserved when the focus is removed and added to same
control. Previously the tests rely on the fact that the
input mode was reset to default when the focus was reset.

Change-Id: I3196919ab4151a629eaa92cbedb21b7d03591051
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>