Fix random fail that occurs in the spell correction tests
This issue occurs randomly when the tests are run in release mode.

This is an actual issue, but may not occur so easily in the real life.

When the input sequence is entered quickly (like in the automated
test script) the candidate list built in the background may override
the last input character. This happens only if the candidate list
build is finished right after the new character is added to the
input word. In this case the old candidate list overrides the exact
match. It is also required that the user selects the outdated result,
for example, by entering a space key, before the latest candidate
list is built.

Change-Id: Ia2d5eb65ea92dc0f159e2b3f822857e63c0f1ec6
Task-number: QTRD-3562
Reviewed-by: Rainer Keller <rainer.keller@theqtcompany.com>