Fix regressions introduced in commit 2b94afe8
In the Keyboard.qml the input mode should always be set when the
input locale changes.

In the hunspell input method, the dictionary should be loaded even
if the prediction is not allowed by the input method hints. It was
originally prevented because the test cases were polluted with
unnecessary warnings by the missing dictionary files. Now the
qWarning is replaced with a debug message and the dictionary is
loaded regardless of the current input method hints.

In the tests, there was an unnecessary Qt.Key_Context1 left in the
japanese test case.

Change-Id: I6442d0c9b485d5126c1bc8b6b7fcd0a240913f0f
Reviewed-by: Rainer Keller <rainer.keller@theqtcompany.com>