Add SVG image provider to Styles module
With SVG image provider, it is possible to request the SVG images
scaled to specified size using the URL parameters.

In normal case, like with the QML Image, this would not be necessary,
since the Image allows setting the sourceSize property, thus allowing
to scale the SVG to desired size.

However, in special cases, such as with BorderImage, the sourceSize is
not available for modification, thus makes it impossible to leverage
SVG scaling capabilities.

With this change, the SVG images can be fetched from the new image
provider available at image://qtvkbsvg URL and with the desired
size given as URL parameters, for example:

"image://qtvkbsvg/path/to/resource/mySvg.svg?width=123&height=456"

It is also possible to give only one dimension, in which case the
other dimension is calculated by the image provider from the design
aspect ratio of the SVG image.

Change-Id: Ided3035d0d414462b2c3eeaba0b8543a4c5ceea9
Reviewed-by: Gatis Paeglis <gatis.paeglis@theqtcompany.com>