Add support for full screen handwriting input
This change adds support for the full screen handwriting input.

In full screen handwriting input, the user is able to use the
whole application screen as a container for handwriting input.

This feature is available to those application integrations,
which integrate the new HandwritingInputPanel type to the
application QML. The HandwritingInputPanel works as an add-on
to the existing InputPanel, so it cannot be used exclusively
without the InputPanel.

Change-Id: I11530b9ccbc66aa59b210bec94d7bb87f0826370
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>