lipi-toolkit: Add new API to shape recognizer for cancelling recognition
The new API allows to asynchronously request cancellation of
the current recognition process.

This API is needed in a use case where the handwriting
recognition process starts as soon as the user stops drawing
the letter, but in which the user can continue drawing before
the results timeout is expired. The new API is used at a
point where the user continues drawing and the ongoing
recognition must be cancelled due to more strokes being
added to character.

The API does not use any synchronization mechanism. However,
it is not strictly needed since the recognizer only reads the
variable used for signaling the cancellation request. Also,
the possible race condition does not cause critical error,
as it would only appear as an unexpected recognition result
from the cancelled operation. Though, it is very unlikely to
happen.

Change-Id: Ibfe7c5f65428614982b60e2dd20ed01af8c3642a
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>