Run lipi HWR task in parallel with the recognition timer
This change improves the user experience by reducing the time
spent waiting the HWR task to complete.

Previously the recognition was started after the recognition
timer expired. This caused the total overhead of the recognition
to be summed on top of the recognition timer, reducing the user
experience in low performance devices. Now the idle time (during
the recognition timer) is used for the benefit of HWR task, so that
in optimal case, the recognition result is available immediately
after the recognition timer expires. If the HWR task takes longer
to process than the recognition timer, the results will be provided
as soon as the HWR task is done.

Ongoing recognition task is cancelled if the user continues drawing
while the recognition timer is running. If the recognition task is
already completed, the current result is ignored and the result
from the next recognition task is used instead.

Change-Id: I9ba797223d8a9b8daf423e500fcf9d5250caaa5a
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>