Add support for gestures (lipi input method)
This change adds support for simple swipe gestures.
The swipe gesture is essential when the handwriting
is used without a keyboard.

The algorithm is experimental, but seems to work
well enough. It supports multi-swipe (e.g. swipe with
two or more fingers).

The following functionality is added for swipe gestures:
- right-to-left single swipe: deletes a character
- right-to-left double swipe: clears the current word, or
  deletes a character
- left-to-right single swipe: inserts a space character
- left-to-right double swipe: commits the current word, or
  inserts a space
- bottom-to-up single swipe: toggles input mode (numbers,
  latin)
- bottom-to-up double swipe: toggles text case.

Change-Id: I0069e4bfdada863854730895fdbd253d43b7fef0
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>