Update input mode when input method hints change
This change fixes an issue that the input mode is not updated when the
input method hints are updated, even though the new input method hints
would trigger different input mode. This will happen if the new hints
do not trigger a keyboard layout switch. Specifically, this applies to
handwriting input, which can be used in combination of all input method
hints.

Change-Id: I942d556ac29b542fe6acc9e2ee35f52f6e12cbc4
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>