Fix input panel visibility
If showInputPanel() is called before setting focus to ImEnabled input
control, the visibility information is lost and the input panel is not
shown until the user presses the input control second time.

Instead, the last visibility state is always stored in the input
context and the input panel visibility is updated according to latest
state.

Change-Id: I479925a7c399068ed0b3cba14aa0b405e590ea6e
Reviewed-by: Rainer Keller <rainer.keller@theqtcompany.com>