Add support for smallText in the BaseKey
This change adds support for smallText label in the key and also
modifies the default behavior of alternative keys.

The smallText property in the BaseKey is by default set to default
alternative key. The keyboard layout can modify the smallText if
necessary. However, the smallText is not visible by default. It can
be set to visible at KeyboardLayout, KeyboardRow, KeyboardColumn or
Key level.

Previously the alternative keys included the "parent" key as the
default highlighted item in the list. This behavior is now changed,
and the "parent" item is filtered out from the effective list presented
to user. Rationale for this is that if the smallText feature would be
turned on for existing layouts, they would render the same key label
twice.

Change-Id: I543b2d3355be640d80f751444944a6dcefbbce73
Reviewed-by: Rainer Keller <rainer.keller@theqtcompany.com>