Fix default locale in a specific case
Reset to default locale if VirtualKeyboardSettings.activeLocales
is updated and VirtualKeyboardSettings.locale is set (to some value).

This fixes a case where the activeLocales is initially set to some
values, which does not include the VirtualKeyboardSettings.locale.
Later, when the activeLocales is updated to include the
VirtualKeyboardSettings.locale, the keyboard layout would not change
immediately to specific locale, as one would expect.

This issue was found by running the test cases with CONFIG+=openwnn
and observing that test_baseKeyNoModifier did not pass.

Change-Id: Iaf3f26a81716a83e936820c5b5f1069916ee81eb
Reviewed-by: Rainer Keller <rainer.keller@theqtcompany.com>