Wrap C++ API into namespace
This change wraps the entire C++ API into namespace.

In practice, all the C++ interfaces inside the qtvirtualkeyboard plugin
are private, except the QPlatformInputContext plugin API. Even the
AbstractInputMethod and AbstractInputPanel are not really a public in
the sense that they could be used outside the plugin. At least it does
not make sense, since there is no way to extend the virtual keyboard
functionality without recompiling the plugin.

Task-number: QTRD-3628
Change-Id: I1037ee247abca3219efeaa4e4150baaff7c3d668
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>