Stabilize arrow key navigation tests
The arrow key navigation tests could fail due to bug in a test function.

Previously the navigation key press and the highlight rectangle could
go out of sync due to highlight animation. Now the test waits until the
highlight animation is completed and then proceeds with the test, making
the test cases more robust and not dependent of system performance.

Change-Id: I72bc15de988a212fa2e96230bc346bbfd3fdea41
Reviewed-by: Rainer Keller <rainer.keller@theqtcompany.com>