Fix crash on vkb when exiting Qt with QCoreApplicaton::quit()
This change fixes an issue that caused crash on Windows desktop
integration when the application was shutdown using the
QCoreApplication::quit() method.

The crash was caused by too late cleanup of the desktop input
panel. Now the virtual keyboard view is closed upon
QCoreApplication::aboutToQuit() signal instead of
QCoreApplication::lastWindowClosed() signal. This seems to work
also in case where the application is shutdown by closing all
the windows.

A workaround for the issue is to connect the
QCoreApplication::lastWindowClosed() signal to
the QCoreApplication::quit() instead of invoking it directly, and then
closing all the windows upon exit. This allows the virtual keyboard
to cleanup its own view before it's too late.

Task-number: QTRD-3747
Change-Id: Ib0915bdab408f85635370205a9504ff68c3a38cb
Reviewed-by: Milla Pohjanheimo <milla.pohjanheimo@theqtcompany.com>
Reviewed-by: Kalle Viironen <kalle.viironen@theqtcompany.com>