Remove AutoScroller component
The AutoScroller component is not well maintained and has not been validated
to be working correctly except for the example app.

This change moves the AutoScroller component to the example application.
There was also a bug in the component. References to the flickable components
were not cleared before updating.

Task-number: QTRD-3784
Change-Id: Ie15b4b994c016a9d0c065a2e469f928fec15d265
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>
Reviewed-by: Rainer Keller <rainer.keller@theqtcompany.com>