Fix language switch for handwriting keyboard layout
Handwriting layout uses ChangeLanguageKey with customLayoutsOnly
attribute to filter out languages which does not have handwriting
layout.

By default the customLayoutsOnly attribute filters out the base layout
(English) if it's not the current language. This functionality exists
for the special layouts, such as the numbers layout, to prevent language
switch between the languages where the layout does not change. However,
for handwriting the input language is significant and it should be
possible to switch between all the languages which contain handwriting
layout.

This change adds an exception for the no-base-layout rule and includes
English handwriting layout in the custom layout list.

Change-Id: I58449c18b20c2643f5cc905163acefe11fd1fa6c
Reviewed-by: Rainer Keller <rainer.keller@theqtcompany.com>