Fix lipi plugin install location
Installing plugins into a data path is a very bad idea, even more so
when the path disagrees with the non-prefix location.

The default LIPI_ROOT contains both the plugins and the projects needed
at runtime for handwriting recognition. This change adds new environment
variable LIPI_LIB which points to location for plugins only. This allows
the separation of the two.

Change-Id: Ie50b6871b6d176b2d06ff068743b40e9daa2c5cb
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>