Implemented proper double click behavior of shift handler
Previously, there is nothing to separate whether the button
command is a single click or a double click. Every time we
click the shift button, it always look for the second click
and always activate the caps lock. There is no time limit to
specify if there will be a second click and always recognized
the upcoming click to activate the caps lock. The correct be-
havior of a shift button is that a single click is used to
enable the upper-case and automatically turned into lower case
when the next letter is typed. Another case, within a specific
time, if a second click comes, the caps-lock is activated and
keep the upper-case mode, otherwise the upcoming click is used
to abort the upper-case and change the keyboard mode back to
lower-case.

Unit test is also updated.

[ChangeLog] Changed behavior of shift handler to only activate
caps lock if the shift key is double-clicked.

Change-Id: Ia04a61ca6df5407f37eb73b9de65c6ccd0128547
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>