Add language popup
This change adds language popup as an alternative method for
selecting the input language.

The language popup is enabled when the active style supports it.
This change adds the support for the default style.

The popup opens from the change language key with single tap
and can be dismissed by tapping anywhere else on the keyboard.

The old toggle method for changing the input language is still
available and supported (can be enabled easily from keyboard style).

Some basic tests are included in this change.

[ChangeLog] Added language selection popup for faster selection of
input language.

Change-Id: Ie3773f1d0cac78dee8237285e8596fe57c8bb5e4
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>