Hunspell: Update selection list status when loading dictionary
HunspellInputMethod must update selection list status immediately
after dictionary starts load.

This issue should not cause issues in normal use, but it will cause
automatic tests to fail with lipi-toolkit. This is because LipiInputMethod
uses HunspellInputMethod for spell correction and suggestions.

The automated tests rely on selection list status when deciding if
spell correction is available or not.

The failing tests was test_hwrWordReselection_data.

Also, fix incorrect debug stating dictionary could not be found.

Change-Id: Ic6c3526a494fb60dde234c000ed9bd0d70f53335
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>