styles.qrc: Remove qmldir
If a plugin is used, the qmldir file must be located next to
the plugin in the file system. By it being included in the .qrc
file it was deployed twice, once in the file system and once
in the virtual resource file system. A recent change in declarative
ended up doing an implicit import on that virtual directory and
thus uncovered this bug of the qmldir file being present in a (virtual)
directory where it doesn't belong.

qml_module.prf takes care of deploying the qmldir.

Task-number: QTBUG-62138
Change-Id: I7693b5656cb593454e2d777fe179e6f889d6dbdf
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>