Fix rendering of high and low acute accent keys in Zhuyin keyboard
The MODIFIER LETTER ACUTE ACCENT and MODIFIER LETTER GRAVE ACCENT were
not rendered correctly on the keyboard layout. Fix it by defining
regular accent characters in the displayText property.

After adding the displayText, it was noticed that the character preview
is not using the BaseKey::displayText for rendering. So fix it too.

Change-Id: If4889729a33a31d6adaf149ca4d263a2004abe7e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>