Prefer current input mode when switching between handwriting mode
If a keyboard layout provides multiple choices for input modes
(e.g. latin and cyrillic), prefer the current choice when switching
to handwriting layout and back.

Also, use the same mechanism to avoid binding loop when switching
input mode with InputModeKey contained in a KeyboardLayoutLoader
where the layout depends on the input mode.

Change-Id: If7adae92db39f3b1e49ec362679e6b6ae7f32fa4
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>