3rdparty/t9write: unpack.py: Allow duplicate file names
This change allows the rules with same input file name (but
with different directory prefix) to co-exist. It works by
removing the already matched input file from the "zip list".
However, this only works if the rules are ordered correctly,
i.e. the most exact ones being first in the list.

Consider the following example:

'data/arabic': [
    '*/Arabic/_databas_le.bin',
],
'data': [
    '*/_databas_le.bin',
]

Without this change the later rule will override the
file entry already copied for 'data/arabic'. But with
'Arabic/_databas_le.bin' removed from the input list, it
will not be matched.

Change-Id: Ifd06125d1519c45dba077e50f69769fdb0070351
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>