Fix QML warning propagated from BaseKey in a specific use case
If the Key.alternativeKeys are assigned like:

Key {
    alternativeKeys: condition ? "abc" : "def"
}

it will cause the following QML warning from BaseKey:

[...]/BaseKey.qml:82:32: Unable to assign [undefined] to QString

It seems to be caused by the order in which the dependent
properties effectiveAlternativeKeysHighlightIndex and
effectiveAlternativeKeys are updated.

This change fixes the issue by adding the effectiveAlternativeKeys
as condition for the property assignment.

Change-Id: I0225a95a60920201161c7d8c084667c57c1c29d4
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>